requirements_opt.txt
# for quantization
optimum-quanto
--extra-index-url https://download.pytorch.org/whl/cu124
torchao